requirements.txt
Pillow==9.5.0
Pillow==11.0.0
rembg==2.0.28
streamlit==1.25.0
streamlit_option_menu==0.3.13
